Add StoreKit config and update IAP product IDs
Added Products.storekit file for StoreKit testing and integrated it into the Xcode project and workspace. Updated in_app_purchase_service.dart to use the correct monthly product ID and added debug logging for initialization and product loading. Yearly product ID is commented out as it is not present in the StoreKit configuration.